this version is major change to make code more clean by splitting the code resolver and typedef in separated file
